+++ lib/model_settings/adapters/column.rb
@@ -40,8 +40,11 @@ module ModelSettings
           # Execute before_enable callback
           execute_setting_callbacks(setting_obj, :enable, :before)
 
-          # Set the value
-          public_send("#{setting_name}=", true)
+          # Execute around_enable callback (wraps the value assignment)
+          execute_around_callback(setting_obj, :enable) do
+            # Set the value
+            public_send("#{setting_name}=", true)
+          end
 
           # Execute after_enable callback
           execute_setting_callbacks(setting_obj, :enable, :after)
@@ -55,8 +58,11 @@ module ModelSettings
           # Execute before_disable callback
           execute_setting_callbacks(setting_obj, :disable, :before)
 
-          # Set the value
-          public_send("#{setting_name}=", false)
+          # Execute around_disable callback (wraps the value assignment)
+          execute_around_callback(setting_obj, :disable) do
+            # Set the value
+            public_send("#{setting_name}=", false)
+          end
 
           # Execute after_disable callback
           execute_setting_callbacks(setting_obj, :disable, :after)
@@ -70,6 +76,8 @@ module ModelSettings
           # Execute before_toggle callback
           execute_setting_callbacks(setting_obj, :toggle, :before)
 
+          # Execute around_toggle callback (wraps the value assignment)
+          # Note: We don't have around_toggle in spec, but keeping consistent pattern
           # Toggle the value
           public_send("#{setting_name}=", !public_send(setting_name))