+++ lib/model_settings/adapters/store_model.rb
@@ -66,8 +66,11 @@ module ModelSettings
           # Execute before_enable callback
           execute_setting_callbacks(setting_obj, :enable, :before)
 
-          # Set the value
-          public_send("#{setting_name}=", true)
+          # Execute around_enable callback (wraps the value assignment)
+          execute_around_callback(setting_obj, :enable) do
+            # Set the value
+            public_send("#{setting_name}=", true)
+          end
 
           # Execute after_enable callback
           execute_setting_callbacks(setting_obj, :enable, :after)
@@ -81,8 +84,11 @@ module ModelSettings
           # Execute before_disable callback
           execute_setting_callbacks(setting_obj, :disable, :before)
 
-          # Set the value
-          public_send("#{setting_name}=", false)
+          # Execute around_disable callback (wraps the value assignment)
+          execute_around_callback(setting_obj, :disable) do
+            # Set the value
+            public_send("#{setting_name}=", false)
+          end
 
           # Execute after_disable callback
           execute_setting_callbacks(setting_obj, :disable, :after)